+++ varfish/users/tasks.py
@@ -3,9 +3,9 @@ import logging
 from celery.schedules import crontab
 from django.conf import settings
 from django.core.management import call_command
+from projectroles.models import RemoteSite, SODAR_CONSTANTS
 
 from config.celery import app
-
 
 #: Logger to use in this module.
 LOGGER = logging.getLogger(__name__)
@@ -14,7 +14,8 @@ LOGGER = logging.getLogger(__name__)
 @app.task(bind=True)
 def sodar_sync_remote(_self):
     try:
-        call_command("syncremote")
+        if RemoteSite.objects.filter(mode=SODAR_CONSTANTS["SITE_MODE_SOURCE"]).count():
+            call_command("syncremote")
     except Exception as e:
         LOGGER.error("Problem with syncremote job: %s", e)