+++ HISTORY.rst
@@ -29,6 +29,7 @@ End-User Summary
 - Fixed broken VariantValidator query (#523).
 - Fixed smallvariant flags filter query (#502).
 - Added flags `segregates`, `doesnt_segregate` and `no_disease_association` to file export (#502).
+- Adding feature to enable and configure link-out to HGMD (#576).
 
 Full Change List
 ================
@@ -53,6 +54,7 @@ Full Change List
 - Fixed smallvariant flags filter query (#502).
 - Added flags `segregates`, `doesnt_segregate` and `no_disease_association` to file export (#502).
 - Converted not cooperative tooltip to standard title on Filter & Display button (#508).
+- Adding feature to enable and configure link-out to HGMD (#576).
 
 ------
 v1.2.0